added view for updating, added task_form.html and update.html

+++ tasks/views.py
@@ -9,6 +9,13 @@ def tasks(request):
     context = {'tasks': tasks}
     return render(request, 'tasks/tasks.html', context)
 
+
+def details(request, *args, pk, **kwargs):
+    task = get_object_or_404(Task, pk=pk)
+    context = {'task': task}
+    return render(request, "tasks/details.html", context)
+
+
 def create_task(request):
     if request.method == 'POST':
         form = TaskForm(request.POST)
@@ -19,12 +26,23 @@ def create_task(request):
         form = TaskForm()
     return render(request, 'tasks/create_task.html', {'form': form})
 
-def delete_task(request, pk):
-    task = Task.objects.get(id=pk)
-    task.delete()
-    return redirect('main')
-
-def details(request, *args, pk, **kwargs):
+def task_update(request, pk, *args, **kwargs):
     task = get_object_or_404(Task, pk=pk)
-    context = {'task': task}
-    return render(request, "tasks/details.html", context)
+    form = TaskForm(instance=task)
+    context = {'form': form}
+
+    if request.method == 'GET':
+        return render(request, 'tasks/task_update.html', context)   #добавить#
+    elif request.method == 'POST':
+        form = TaskForm(request.POST)
+        if form.is_valid():
+            task = form.save()
+            return redirect("detail", pk=task.pk)
+        return render(request, 'tasks/task_update.html', {'form': form})
+
+
+def delete_task(request, pk):
+    if request.method == 'POST':
+        task = get_object_or_404(Task, pk=pk)
+        task.delete()
+    return redirect('main')
